test: 💍 move HomeScreenUiTest to test folder
